refactor: update references to match, rename `graphql/soil_id` files
